middlewares and requests TS migration
