+++ src/app/http/requests/forgot-password-store.request.ts
@@ -1,8 +1,14 @@
+import { Request, Response, NextFunction } from 'express';
+
 const { body, validationResult } = require('express-validator');
 
 const forgotPasswordRequest = [
-  body('email').isLength({ min: 1, max: 255 }).trim().isEmail().normalizeEmail({ gmail_remove_dots: false }),
-  (req, res, next) => {
+  body('email')
+    .isLength({ min: 1, max: 255 })
+    .trim()
+    .isEmail()
+    .normalizeEmail({ gmail_remove_dots: false }),
+  (req: Request, res: Response, next: NextFunction) => {
     const errors = validationResult(req);
 
     if (!errors.isEmpty()) {
